Rename and re-arrange files
