+++ resources/js/src/components/Label.js
